src_backend/pyproject.toml
[tool.poetry]
name = "my-site"
version = "0.1.0"
description = ""
authors = ["Lunev Dmitriy <dima.lunev14@gmail.com>"]
readme = "README.md"
# packages = [{include = "my_site"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.88.0"}
typer = {extras = ["all"], version = "^0.7.0"}
pytest = "^7.2.0"
loguru = "^0.6.0"
tortoise-orm = {extras = ["asyncpg"], version = "^0.19.2"}
aerich = "^0.7.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^5.0.4"
flake8-commas = "^2.1.0"
flake8-docstrings = "^1.6.0"
flake8-isort = "^5.0.0"
flake8-encodings = "^0.5.0.post1"
flake8-secure-coding-standard = "^1.3.0"
flake8-string-format = "^0.3.0"
darglint = "^1.8.1"
flake8-comprehensions = "^3.10.1"
tryceratops = "^1.1.0"
flake8-datetimez = "^20.10.0"
flake8-implicit-str-concat = "^0.3.0"
flake8-return = "^1.2.0"
flake8-scream = "^0.1.0"
flake8-too-many = "^0.1.5"
flake8-broken-line = "^0.6.0"
flake8-pep3101 = "^2.0.0"
flake8-print = "^5.0.0"
flake8-use-pathlib = "^0.3.0"
flake8-functions-names = "^0.3.0"
flake8-variables-names = "^0.0.5"
pep8-naming = "^0.13.2"
flake8-fastapi = "^0.7.0"
tortoise-orm-stubs = "^0.3.7"
flake8-bugbear = "^22.10.27"
coverage = {extras = ["toml"], version = "^7.0.3"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.aerich]
tortoise_orm = "app.core.config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."

[tool.black]
line-length = 100
skip-string-normalization = true

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3

[tool.coverage]
    [tool.coverage.run]

    source = ["app"]
    branch = true
    omit = ['']

    [tool.coverage.report]
    exclude_lines = [
        "# noqa",
        "raise NotImplementedError"
    ]

[tool.pytest]
testpath = "tests"
